+++ src/SMAPI/Framework/ContentCoordinator.cs
@@ -32,6 +32,9 @@ namespace StardewModdingAPI.Framework
         /// <summary>An asset key prefix for assets from SMAPI mod folders.</summary>
         private readonly string ManagedPrefix = "SMAPI";
 
+        /// <summary>Whether to use raw image data when possible, instead of initializing an XNA Texture2D instance through the GPU.</summary>
+        private readonly bool UseRawImageLoading;
+
         /// <summary>Get a file lookup for the given directory.</summary>
         private readonly Func<string, IFileLookup> GetFileLookup;
 
@@ -130,7 +133,8 @@ namespace StardewModdingAPI.Framework
         /// <param name="getFileLookup">Get a file lookup for the given directory.</param>
         /// <param name="onAssetsInvalidated">A callback to invoke when any asset names have been invalidated from the cache.</param>
         /// <param name="requestAssetOperations">Get the load/edit operations to apply to an asset by querying registered <see cref="IContentEvents.AssetRequested"/> event handlers.</param>
-        public ContentCoordinator(IServiceProvider serviceProvider, string rootDirectory, CultureInfo currentCulture, IMonitor monitor, Reflector reflection, JsonHelper jsonHelper, Action onLoadingFirstAsset, Action<BaseContentManager, IAssetName> onAssetLoaded, Func<string, IFileLookup> getFileLookup, Action<IList<IAssetName>> onAssetsInvalidated, Func<IAssetInfo, AssetOperationGroup?> requestAssetOperations)
+        /// <param name="useRawImageLoading">Whether to use raw image data when possible, instead of initializing an XNA Texture2D instance through the GPU.</param>
+        public ContentCoordinator(IServiceProvider serviceProvider, string rootDirectory, CultureInfo currentCulture, IMonitor monitor, Reflector reflection, JsonHelper jsonHelper, Action onLoadingFirstAsset, Action<BaseContentManager, IAssetName> onAssetLoaded, Func<string, IFileLookup> getFileLookup, Action<IList<IAssetName>> onAssetsInvalidated, Func<IAssetInfo, AssetOperationGroup?> requestAssetOperations, bool useRawImageLoading)
         {
             this.GetFileLookup = getFileLookup;
             this.Monitor = monitor ?? throw new ArgumentNullException(nameof(monitor));
@@ -141,6 +145,7 @@ namespace StardewModdingAPI.Framework
             this.OnAssetsInvalidated = onAssetsInvalidated;
             this.RequestAssetOperations = requestAssetOperations;
             this.FullRootDirectory = Path.Combine(Constants.GamePath, rootDirectory);
+            this.UseRawImageLoading = useRawImageLoading;
             this.ContentManagers.Add(
                 this.MainContentManager = new GameContentManager(
                     name: "Game1.content",
@@ -219,7 +224,8 @@ namespace StardewModdingAPI.Framework
                     reflection: this.Reflection,
                     jsonHelper: this.JsonHelper,
                     onDisposing: this.OnDisposing,
-                    fileLookup: this.GetFileLookup(rootDirectory)
+                    fileLookup: this.GetFileLookup(rootDirectory),
+                    useRawImageLoading: this.UseRawImageLoading
                 );
                 this.ContentManagers.Add(manager);
                 return manager;
@@ -465,7 +471,7 @@ namespace StardewModdingAPI.Framework
                     assets: invalidatedAssets.ToDictionary(p => p.Key, p => p.Value),
                     ignoreWorld: Context.IsWorldFullyUnloaded,
                     out IDictionary<IAssetName, bool> propagated,
-                    out bool updatedNpcWarps
+                    out bool updatedWarpRoutes
                 );
 
                 // log summary
@@ -481,8 +487,8 @@ namespace StardewModdingAPI.Framework
                         ? $"Propagated {propagatedKeys.Length} core assets ({FormatKeyList(propagatedKeys)})."
                         : "Propagated 0 core assets."
                     );
-                    if (updatedNpcWarps)
-                        report.AppendLine("Updated NPC pathfinding cache.");
+                    if (updatedWarpRoutes)
+                        report.AppendLine("Updated NPC warp route cache.");
                 }
                 this.Monitor.Log(report.ToString().TrimEnd());
             }