+++ src/SMAPI/Properties/AssemblyInfo.cs
@@ -3,3 +3,4 @@ using System.Runtime.CompilerServices;
 [assembly: InternalsVisibleTo("SMAPI.Tests")]
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")] // Moq for unit testing
 [assembly: InternalsVisibleTo("ContentPatcher")]
+[assembly: InternalsVisibleTo("ErrorHandler")]