+++ docs/release-notes.md
@@ -1,11 +1,31 @@
 ← [README](README.md)
 
 # Release notes
+## 3.15.0
+Released 17 June 2022 for Stardew Valley 1.5.6 or later. See [release highlights](https://www.patreon.com/posts/67877219).
+
+* For players:
+  * Optimized mod image file loading.
+  * Minor optimizations (thanks to Michael Kuklinski / Ameisen!).
+  * Updated compatibility list.
+
+* For mod authors:
+  * Added an [`IRawTextureData` asset type](https://stardewvalleywiki.com/Modding:Migrate_to_SMAPI_4.0#Raw_texture_data), to avoid creating full `Texture2D` instances in many cases.
+  * In `smapi-internal/config.json`, you can now enable verbose logging for specific mods (instead of all or nothing).
+  * Updated dependencies:
+    * Harmony 2.2.1 (see changes in [2.2.0](https://github.com/pardeike/Harmony/releases/tag/v2.2.0.0) and [2.2.1](https://github.com/pardeike/Harmony/releases/tag/v2.2.1.0));
+    * Newtonsoft.Json 13.0.1 (see [changes](https://github.com/JamesNK/Newtonsoft.Json/releases/tag/13.0.1));
+    * Pintail 2.2.0.
+  * Removed transitional `UsePintail` option added in 3.14.0 (now always enabled).
+  * Fixed `onBehalfOf` arguments in the new content API being case-sensitive.
+  * Fixed map edits which change warps sometimes rebuilding the NPC pathfinding cache unnecessarily, which could cause a noticeable delay for players.
+
 ## 3.14.7
 Released 01 June 2022 for Stardew Valley 1.5.6 or later.
 
 * For players:
   * Optimized reflection cache to reduce frame skips for some players.
+
 * For mod authors:
   * Removed `runtimeconfig.json` setting which impacted hot reload support.